backend/requirements.txt
annotated-types==0.6.0
anyio==4.3.0
certifi==2024.2.2
click==8.1.7
dnspython==2.6.1
email_validator==2.1.1
fastapi==0.110.0
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
idna==3.6
iniconfig==2.0.0
motor==3.3.2
packaging==24.0
pluggy==1.4.0
pydantic==2.6.3
pydantic_core==2.16.3
pymongo==4.6.2
pytest==8.1.1
pytest-asyncio==0.23.6
python-dotenv==1.0.1
sniffio==1.3.1
starlette==0.36.3
typing_extensions==4.10.0
uvicorn==0.28.0
